Se agregada base de datos , postgresql correciones en entidades y properties, se agrega docker-compose.yaml

+++ src/main/java/com/devsu/api/bancario/entity/PersonaEntity.java
@@ -33,7 +33,7 @@ public class PersonaEntity implements Serializable {
 
     @NotNull(message = "Edad debe ser ingresado")
     @Range(min=1, max=150,message = "Edad minima 1, edad maxima 150")
-    private byte edad;
+    private Integer edad;
 
     @NotEmpty(message = "Identificacion debe ser ingresado")
     @Column(unique = true)
